marathon-api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
scikit-learn==1.4.0
python-dotenv==1.0.1
joblib==1.3.2
numpy>=1.24.0
pydantic==2.6.1
requests==2.31.0